# restler-fuzzer-main/Compile/preprocessed/swagger_preprocessed.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/realtor/create":{"post":{"tags":["hpms-endpoints"],"summary":"Create a realtor record","description":"Create a realtor record","operationId":"createRealtor","parameters":[{"name":"employeenum","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"firstName","in":"query","description":"First name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]+","type":"string","description":"First name containing only letters"}},{"name":"lastName","in":"query","description":"Last name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]+","type":"string","description":"Last name containing only letters"}},{"name":"commissionRate","in":"query","description":"Commission Rate in float format (min: 0, max: 1)","required":true,"schema":{"maximum":1.0,"exclusiveMaximum":false,"minimum":0.0,"exclusiveMinimum":false,"type":"number","description":"Commission Rate in float format (min: 0, max: 1)","format":"float"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RealtorId"}}}}}}},"/purchase/create":{"post":{"tags":["hpms-endpoints"],"summary":"Create a purchase record","description":"Create a purchase record","operationId":"createPurchase","parameters":[{"name":"customerId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"realtorId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"homeId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"loan","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"downPayment","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"404":{"description":"Resource not possible","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PurchaseId"}}}}}}},"/home/create":{"post":{"tags":["hpms-endpoints"],"summary":"Create a home record","description":"Create a home record","operationId":"createHome","parameters":[{"name":"address","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"street","in":"query","description":"Street name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]+","type":"string","description":"Street name containing only letters"}},{"name":"city","in":"query","description":"City name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]+","type":"string","description":"City name containing only letters"}},{"name":"state","in":"query","description":"State name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]{2,4}","type":"string","description":"State name containing only letters"}},{"name":"price","in":"query","description":"Price in integer format (min: 1, max: 9999999)","required":true,"schema":{"maximum":9999999.0,"exclusiveMaximum":false,"minimum":1.0,"exclusiveMinimum":false,"type":"number","description":"Price in integer format (min: 1, max: 9999999)","format":"float"}},{"name":"squareFeet","in":"query","description":"Square feet in integer format (min: 1)","required":true,"schema":{"minimum":1.0,"exclusiveMinimum":false,"type":"integer","description":"Square feet in integer format (min: 1)","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HomeId"}}}}}}},"/customer/create":{"post":{"tags":["hpms-endpoints"],"summary":"Create a customer record","description":"Create a customer record","operationId":"createCustomer","parameters":[{"name":"firstName","in":"query","description":"First name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]+","type":"string","description":"First name containing only letters"}},{"name":"lastName","in":"query","description":"Last name containing only letters","required":true,"schema":{"pattern":"[a-zA-Z]+","type":"string","description":"Last name containing only letters"}},{"name":"ssn","in":"query","description":"Social Security Number format XX-XXX-XXXX","required":true,"schema":{"pattern":"[0-9]{2}-[0-9]{3}-[0-9]{4}","type":"string","description":"Social Security Number format XX-XXX-XXXX"}},{"name":"income","in":"query","description":"Annual income in float format (min: 0, max: 1,0000,000)","required":true,"schema":{"maximum":10000000,"exclusiveMaximum":false,"minimum":0.0,"exclusiveMinimum":false,"type":"number","description":"Annual income in float format (min: 0, max: 1,0000,000)","format":"float"}}],"responses":{"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerId"}}}},"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}}}}},"/realtor/get/{realtorid}":{"get":{"tags":["hpms-endpoints"],"summary":"Get a realtor record","description":"Get a realtor record","operationId":"getRealtor","parameters":[{"name":"realtorid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Realtor"}}}}}}},"/purchase/get/{purchaseid}":{"get":{"tags":["hpms-endpoints"],"summary":"Create a purchase record","description":"Create a purchase record","operationId":"getPurchase","parameters":[{"name":"purchaseid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Purchase"}}}}}}},"/home/get/{homeid}":{"get":{"tags":["hpms-endpoints"],"summary":"Get a home record","description":"Get a home record","operationId":"getHome","parameters":[{"name":"homeid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Home"}}}}}}},"/customer/get/{customerid}":{"get":{"tags":["hpms-endpoints"],"summary":"Create a purchase record","description":"Create a purchase record","operationId":"getCustomer","parameters":[{"name":"customerid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"200":{"description":"Resource Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Customer"}}}}}}},"/realtor/delete/{realtorid}":{"delete":{"tags":["hpms-endpoints"],"summary":"Delete a realtor record","description":"Delete a realtor record","operationId":"deleteRealtor","parameters":[{"name":"realtorid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"204":{"description":"Resource Deleted"}}}},"/purchase/delete/{purchaseid}":{"delete":{"tags":["hpms-endpoints"],"summary":"Delete a purchase record","description":"Delete a purchase record","operationId":"deletePurchase","parameters":[{"name":"purchaseid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"204":{"description":"Resource Deleted"}}}},"/home/delete/{homeid}":{"delete":{"tags":["hpms-endpoints"],"summary":"Delete a home record","description":"Delete a home record","operationId":"deletehome","parameters":[{"name":"homeid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"204":{"description":"Resource Deleted"}}}},"/customer/delete/{customerid}":{"delete":{"tags":["hpms-endpoints"],"summary":"Delete a customer record","description":"Delete a customer record","operationId":"deleteCustomer","parameters":[{"name":"customerid","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"400":{"description":"Invalid request parameters","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BadRequestErrorSchema"}}}},"404":{"description":"Resource not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotFoundErrorSchema"}}}},"503":{"description":"Service Unavailable","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ServiceUnavailableErrorSchema"}}}},"204":{"description":"Resource Deleted"}}}}},"components":{"schemas":{"BadRequestErrorSchema":{"type":"object","properties":{"error":{"type":"string","description":"Error type","example":"Bad Request"},"message":{"type":"string","description":"Detailed message","example":"One or more inputs invalid"},"status":{"type":"integer","description":"HTTP status code","format":"int32","example":400},"timestamp":{"type":"string","description":"Timestamp of the error","format":"date-time"}},"description":"Service Unavailable response details"},"ServiceUnavailableErrorSchema":{"type":"object","properties":{"error":{"type":"string","description":"Error type","example":"Service Unavailable"},"message":{"type":"string","description":"Detailed message","example":"Service Unavailable: Issue with server"},"status":{"type":"integer","description":"HTTP status code","format":"int32","example":503},"timestamp":{"type":"string","description":"Timestamp of the error","format":"date-time"}},"description":"Service Unavailable response details"},"RealtorId":{"type":"object","properties":{"realtorid":{"type":"integer","format":"int32"}}},"NotFoundErrorSchema":{"type":"object","properties":{"error":{"type":"string","description":"Error type","example":"Not Found"},"message":{"type":"string","description":"Detailed message","example":"Resource not found"},"status":{"type":"integer","description":"HTTP status code","format":"int32","example":404},"timestamp":{"type":"string","description":"Timestamp of the error","format":"date-time"}},"description":"Error response details"},"PurchaseId":{"type":"object","properties":{"purchaseid":{"type":"integer","format":"int32"}}},"HomeId":{"type":"object","properties":{"homeid":{"type":"integer","format":"int32"}}},"CustomerId":{"type":"object","properties":{"customerid":{"type":"integer","format":"int32"}}},"Realtor":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"employeenum":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"commissionRate":{"type":"number","format":"float"}}},"Purchase":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"customerid":{"type":"integer","format":"int32"},"realtorid":{"type":"integer","format":"int32"},"homeid":{"type":"integer","format":"int32"},"loan":{"type":"integer","format":"int32"},"downpayment":{"type":"integer","format":"int32"}}},"Home":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"address":{"type":"integer","format":"int32"},"street":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"price":{"type":"number","format":"float"},"squareFeet":{"type":"integer","format":"int32"}}},"Customer":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"firstName":{"type":"string"},"lastName":{"type":"string"},"ssn":{"type":"string"},"income":{"type":"number","format":"float"}}}}}}