# src/main/java/com/CSI5370/HomePurchaseManagementSystem/Domain/CustomerId.java
package com.CSI5370.HomePurchaseManagementSystem.Domain;


import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
public class CustomerId {
    public int customerid;
}