# src/main/java/com/CSI5370/HomePurchaseManagementSystem/Domain/HomeId.java
package com.CSI5370.HomePurchaseManagementSystem.Domain;


import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
public class HomeId {
    public int homeid;
}