# src/main/java/com/CSI5370/HomePurchaseManagementSystem/Domain/PurchaseId.java
package com.CSI5370.HomePurchaseManagementSystem.Domain;


import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
public class PurchaseId {
    public int purchaseid;
}