# src/main/java/com/CSI5370/HomePurchaseManagementSystem/Domain/RealtorId.java
package com.CSI5370.HomePurchaseManagementSystem.Domain;


import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
public class RealtorId {
    public int realtorid;
}